moved photometry to utils
