--- threeML/plugins/photometry/__init__.py
@@ -1,1 +1,0 @@
-#from .filter_factory import threeML_filter_library
